refactor: validation 위치 변경
